+++ src/main/java/model/E.java
@@ -3,7 +3,7 @@ package model;
 import java.util.SplittableRandom;
 
 public class E {
-    public static final double MAX_DAMAGE = 7.0;
+    public static final int MAX_DAMAGE = 7;
     public static final SplittableRandom random = new SplittableRandom();
 
     public static final int PLAYER_ONE_ID = 0;
@@ -19,5 +19,5 @@ public class E {
     public static final String RIGHT = "RIGHT";
     public static final String DOWN = "DOWN";
     public static final String LEFT = "LEFT";
-    public static final double DAMAGE_REDUCTION_COEFF = 1.0;
+    public static final int DAMAGE_REDUCTION_COEFF = 1;
 }
